+++ fiboa_cli/datasets/ai4sf.py
@@ -111,7 +111,7 @@ Multi-temporal composites of Sentinel-2 (S2) images are provided to ensure cloud
         gdf["fiboa_id"] = (
             gdf["id"].astype(str).str.zfill(2) + "_" + gdf.index.astype(str).str.zfill(5)
         )
-        return gdf
+        return super().migrate(gdf)
 
     missing_schemas = {
         "properties": {