+++ fiboa_cli/datasets/commons/hcat.py
@@ -3,6 +3,8 @@ from io import StringIO
 from typing import Optional
 
 import geopandas as gpd
+import numpy as np
+import pandas as pd
 from vecorel_cli.vecorel.util import load_file
 
 HCAT_EXTENSION = "https://fiboa.org/hcat-extension/v0.3.0/schema.yaml"
@@ -17,7 +19,7 @@ class AddHCATMixin:
 
     ec_mapping_csv: Optional[str] = None  # TODO rename to hcat_mapping_csv
     mapping_file = None
-    ec_mapping: Optional[dict] = None  # TODO rename to hcat_mapping
+    ec_mapping: Optional[list[dict]] = None  # TODO rename to hcat_mapping
 
     hcat_columns = {
         "hcat:name_en": "hcat:name_en",
@@ -38,11 +40,11 @@ class AddHCATMixin:
             )
         return super().convert(*args, **kwargs)
 
-    def get_code_column(self, gdf):
+    def get_code_column(self, gdf, code="crop:code"):
         try:
-            attribute = next(k for k, v in self.columns.items() if v == "crop:code")
+            attribute = next(k for k, v in self.columns.items() if v == code)
         except StopIteration:
-            raise Exception(f"Misssing crop:code column in converter {self.__class__.__name__}")
+            raise Exception(f"Misssing {code} column in converter {self.__class__.__name__}")
         col = gdf[attribute]
         # Should be corrected in original parser
         return col if col.dtype == "object" else col.astype(str)
@@ -56,15 +58,42 @@ class AddHCATMixin:
 
             if self.ec_mapping is None:
                 self.ec_mapping = load_ec_mapping(self.ec_mapping_csv, url=self.mapping_file)
-            crop_code_col = self.get_code_column(gdf)
+
+            from_code = "original_code"
+            if from_code not in self.ec_mapping[0]:
+                # Some code lists have no code, only a crop_name
+                from_code = "original_name"
+                crop_code_col = self.get_code_column(gdf, "crop:name")
+            else:
+                crop_code_col = self.get_code_column(gdf)
 
             def map_to(attribute):
-                return {e["original_code"]: e[attribute] for e in self.ec_mapping}
+                return {e[from_code]: e[attribute] or None for e in self.ec_mapping}
 
+            col = None
             for k, v in zip(
                 self.hcat_columns.keys(), ("translated_name", "HCAT3_name", "HCAT3_code")
             ):
-                gdf[k] = crop_code_col.map(map_to(v))
+                if v in self.ec_mapping[0]:
+                    col = crop_code_col.map(map_to(v))
+                    gdf[k] = col
+                    assert np.unique(col[~col.isna()]).size > 1, "No HCAT crops mapped"
+
+            if col is not None and col.isna().any():
+                index = [
+                    k for k, v in self.columns.items() if v.startswith("crop:") and k in gdf.columns
+                ]
+                missing = gdf[col.isna()][index].drop_duplicates()
+                missing.reset_index(drop=True, inplace=True)
+                with pd.option_context(
+                    "display.max_colwidth",
+                    None,
+                    "display.max_columns",
+                    None,
+                    "display.max_rows",
+                    None,
+                ):
+                    self.info(f"Missing codes in HCAT mapping:\n{missing}")
 
         if "crop:code_list" not in gdf.columns:
             gdf["crop:code_list"] = (
@@ -72,12 +101,14 @@ class AddHCATMixin:
             )
         return gdf
 
-    def migrate(self, gdf) -> gpd.GeoDataFrame:
-        gdf = super().migrate(gdf)
+    def post_migrate(self, gdf) -> gpd.GeoDataFrame:
+        gdf = super().post_migrate(gdf)
         return self.add_hcat(gdf)
 
 
 def ec_url(csv_file):
+    if csv_file.startswith("https://"):
+        return csv_file
     return f"https://raw.githubusercontent.com/maja601/EuroCrops/refs/heads/main/csvs/country_mappings/{csv_file}"