+++ fiboa_cli/datasets/hr.py
@@ -1,11 +1,10 @@
 from vecorel_cli.conversion.admin import AdminConverterMixin
 
 from ..conversion.fiboa_converter import FiboaBaseConverter
-from .commons.data import read_data_csv
-from .commons.ec import ec_url
+from .commons.hcat import AddHCATMixin
 
 
-class Converter(AdminConverterMixin, FiboaBaseConverter):
+class Converter(AdminConverterMixin, AddHCATMixin, FiboaBaseConverter):
     sources = "https://www.apprrr.hr/wp-content/uploads/nipp/land_parcels.gpkg"
     id = "hr"
     short_name = "Croatia"
@@ -28,17 +27,17 @@ and supporting sustainable land use practices.
     license = "Prostorni podaci i servisi <https://www.apprrr.hr/prostorni-podaci-servisi/>"
     index_as_id = True
 
+    column_migrations = {"land_use_id": lambda col: col.astype(int)}
+
     columns = {
         "id": "id",
         "land_use_id": "crop:code",
-        "crop:name": "crop:name",
-        "crop:name_en": "crop:name_en",
         "area": "metrics:area",
         "geometry": "geometry",
         "home_name": "home_name",
         "perim": "metrics:perimeter",
         "slope": "slope",
-        "z_avg": "z_avg",
+        "z_avg": "height",
         "eligibility_coef": "eligibility_coef",
         "mines_status": "mines_status",
         "mines_year_removed": "mines_year_removed",
@@ -60,8 +59,7 @@ and supporting sustainable land use practices.
         "jpaid": "jpaid",
     }
 
-    extensions = {"https://fiboa.org/crop-extension/v0.2.0/schema.yaml"}
-    column_additions = {"crop:code_list": ec_url("hr_2020.csv")}
+    ec_mapping_csv = "hr_2020.csv"
 
     missing_schemas = {
         "required": [
@@ -73,10 +71,10 @@ and supporting sustainable land use practices.
             "jpaid",
         ],
         "properties": {
-            "land_use_id": {"type": "double"},
+            "land_use_id": {"type": "integer"},
             "home_name": {"type": "string"},
             "slope": {"type": "double"},
-            "z_avg": {"type": "double"},
+            "height": {"type": "double"},
             "eligibility_coef": {"type": "double"},
             "mines_status": {"type": "string", "enum": ["N", "M", "R"]},
             "mines_year_removed": {"type": "int32"},
@@ -98,15 +96,6 @@ and supporting sustainable land use practices.
             "jpaid": {"type": "string"},
         },
     }
-
-    def migrate(self, gdf):
-        gdf = super().migrate(gdf)
-        rows = read_data_csv("hr_categories.csv", delimiter=";")
-        mapping = {int(row["code"]): row["name"] for row in rows}
-        mapping_en = {int(row["code"]): row["name_en"] for row in rows}
-        gdf["crop:name"] = gdf["land_use_id"].map(mapping)
-        gdf["crop:name_en"] = gdf["land_use_id"].map(mapping_en)
-        return gdf
 
     area_is_in_ha = False
     area_calculate_missing = True