+++ tests/test_convert.py
@@ -16,18 +16,19 @@ Optionally use `-lco ENCODING=UTF-8` if you have character encoding issues.
 
 tests = [
     "at",
-    "at_crop",
+    "at_block",
     "be_vlg",
     "br_ba_lem",
     "bg",
     "de_sh",
+    "de_bb",
     "ec_lv",
     "ec_si",
     "fi",
     "fr",
     "hr",
     "nl",
-    "nl_crop",
+    "nl_block",
     "pt",
     "dk",
     "be_wal",
@@ -36,6 +37,7 @@ tests = [
     "ch",
     "cz",
     "us_usda_cropland",
+    "us_ca_scm",
     "jp",
     "lv",
     "ie",
@@ -57,7 +59,9 @@ def _input_files(converter, *names):
 
 extra_convert_parameters = {
     "ai4sf": _input_files("ai4sf", "1_vietnam_areas.gpkg", "4_cambodia_areas.gpkg"),
-    "nl_crop": {"variant": "2023"},
+    "nl": {"variant": "2023"},
+    "se": {"variant": "2023"},
+    "si": {"variant": "2023"},
     "be_vlg": {"variant": "2023"},
     "br_ba_lem": _input_files("br_ba_lem", "LEM_dataset.zip"),
     "ch": _input_files("ch", "lwb_nutzungsflaechen_v2_0_lv95.gpkg"),
@@ -74,6 +78,8 @@ def test_converter(load_ec_mock, capsys, tmp_parquet_file, converter):
     from fiboa_cli import Registry  # noqa
 
     def load_ec(csv_file=None, url=None):
+        if csv_file and "://" in csv_file:
+            csv_file = csv_file.split("/")[-1]
         path = url if url and "://" not in url else f"{test_path}/{converter}/{csv_file}"
         return list(DictReader(open(path, "r", encoding="utf-8")))
 
@@ -95,6 +101,6 @@ def test_converter(load_ec_mock, capsys, tmp_parquet_file, converter):
     ValidateData().validate(tmp_parquet_file)
 
     df = pq.read_table(tmp_parquet_file).to_pandas()
-    if "metrics:area" in df.columns:
+    if "metrics:area" in df.columns and converter not in ("de_bb",):
         # Check for accidental hectare conversion; fields should be more than 10 square meters
         assert (df["metrics:area"] > 10).all()